pm/rm/utils: fix MISRA.ETYPE.ASSIGN.2012
The value of an expression shall not be assigned
to an object with a narrower essential type or
of a different essential type category.
Fixed by having the required datatype.

sbool clk_set_parent - function is modified:
second parameter new_parent size changed to u16 -> u8,value of
new_parent get upadted by idx varible which size is 6-bit olny.

Fixes :SYSFW-6414

+++ pm/include/clk_mux.h
@@ -65,6 +65,6 @@ extern const struct clk_drv_mux clk_drv_mux_reg_ro;
 extern const struct clk_drv_mux clk_drv_mux_reg;
 
 const struct clk_parent *clk_get_parent(struct clk *clkp);
-sbool clk_set_parent(struct clk *clkp, u16 new_parent);
+sbool clk_set_parent(struct clk *clkp, u8 new_parent);
 
 #endif
